+++ aws-datastore/src/androidTest/java/com/amplifyframework/datastore/storage/sqlite/ModelUpgradeSQLiteInstrumentedTest.java
@@ -21,7 +21,7 @@ import androidx.test.core.app.ApplicationProvider;
 import com.amplifyframework.AmplifyException;
 import com.amplifyframework.core.Consumer;
 import com.amplifyframework.core.model.ModelSchema;
-import com.amplifyframework.core.model.ModelSchemaRegistry;
+import com.amplifyframework.core.model.SchemaRegistry;
 import com.amplifyframework.datastore.DataStoreException;
 import com.amplifyframework.datastore.StrictMode;
 import com.amplifyframework.datastore.model.CompoundModelProvider;
@@ -37,6 +37,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.util.List;
+import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
 import static org.junit.Assert.assertEquals;
@@ -52,6 +53,7 @@ public final class ModelUpgradeSQLiteInstrumentedTest {
     private SQLiteStorageAdapter sqliteStorageAdapter;
     private AmplifyCliGeneratedModelProvider modelProvider;
     private RandomVersionModelProvider modelProviderThatUpgradesVersion;
+    private SchemaRegistry schemaRegistry;
 
     private Context context;
 
@@ -65,14 +67,20 @@ public final class ModelUpgradeSQLiteInstrumentedTest {
 
     /**
      * Setup the required information for SQLiteStorageHelper construction.
+     *
+     * @throws AmplifyException may throw {@link AmplifyException} from {@link SchemaRegistry#register(Set)}
      */
     @Before
-    public void setUp() {
+    public void setUp() throws AmplifyException {
         context = ApplicationProvider.getApplicationContext();
         context.deleteDatabase(DATABASE_NAME);
 
         modelProvider = AmplifyCliGeneratedModelProvider.singletonInstance();
         modelProviderThatUpgradesVersion = RandomVersionModelProvider.singletonInstance();
+
+        schemaRegistry = SchemaRegistry.instance();
+        schemaRegistry.clear();
+        schemaRegistry.register(modelProvider.models());
     }
 
     /**
@@ -83,6 +91,7 @@ public final class ModelUpgradeSQLiteInstrumentedTest {
     public void tearDown() throws DataStoreException {
         sqliteStorageAdapter.terminate();
         context.deleteDatabase(DATABASE_NAME);
+        schemaRegistry.clear();
     }
 
     /**
@@ -93,10 +102,7 @@ public final class ModelUpgradeSQLiteInstrumentedTest {
     @Test
     public void modelVersionStoredCorrectlyBeforeAndAfterUpgrade() throws AmplifyException {
         // Initialize StorageAdapter with models
-        ModelSchemaRegistry modelSchemaRegistry = ModelSchemaRegistry.instance();
-        modelSchemaRegistry.clear();
-        modelSchemaRegistry.register(modelProvider.models());
-        sqliteStorageAdapter = SQLiteStorageAdapter.forModels(modelSchemaRegistry, modelProvider);
+        sqliteStorageAdapter = SQLiteStorageAdapter.forModels(schemaRegistry, modelProvider);
         List<ModelSchema> firstResults = Await.result(
             SQLITE_OPERATION_TIMEOUT_MS,
             (Consumer<List<ModelSchema>> onResult, Consumer<DataStoreException> onError) ->
@@ -124,7 +130,7 @@ public final class ModelUpgradeSQLiteInstrumentedTest {
         sqliteStorageAdapter = null;
 
         sqliteStorageAdapter =
-            SQLiteStorageAdapter.forModels(modelSchemaRegistry, modelProviderThatUpgradesVersion);
+            SQLiteStorageAdapter.forModels(schemaRegistry, modelProviderThatUpgradesVersion);
 
         // Now, initialize storage adapter with the new models
         List<ModelSchema> secondResults = Await.result(