+++ aws-datastore/src/main/java/com/amplifyframework/datastore/appsync/AppSyncClient.java
@@ -119,9 +119,8 @@ public final class AppSyncClient implements AppSync {
         };
         final Consumer<ApiException> failureConsumer =
             failure -> onFailure.accept(new DataStoreException(
-                "Failure performing sync query to AppSync.",
-                failure, AmplifyException.TODO_RECOVERY_SUGGESTION
-            ));
+                        "Failure performing sync query to AppSync.",
+                        failure, AmplifyException.TODO_RECOVERY_SUGGESTION));
 
         final Cancelable cancelable = api.query(request, responseConsumer, failureConsumer);
         if (cancelable != null) {